- Putting old rendering core back

+++ Source/Core.cuh
@@ -7,6 +7,6 @@
 
 class CScene;
 
-extern "C" void BindDensityVolume(float* pDensityBuffer, cudaExtent Size);
+extern "C" void BindDensityVolume(short* pDensityBuffer, cudaExtent Size);
 extern "C" void BindExtinctionVolume(float* pExtinctionBuffer, cudaExtent Size);
 extern "C" void Render(const int& Type, CScene* pScene, CScene* pDevScene, unsigned int* pSeeds, CColorXyz* pDevEstFrameXyz, CColorXyz* pDevEstFrameBlurXyz, CColorXyz* pDevAccEstXyz, unsigned char* pDevEstRgbLdr, int N);
